+++ npm/vite-dev-server/index.d.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference
-/// <reference path="../../cli/types/cypress.d.ts" />
+export * from "./dist"