+++ packages/net-stubbing/lib/internal-types.ts
@@ -1,8 +1,9 @@
 import * as _ from 'lodash'
 import {
   RouteMatcherOptionsGeneric,
-  CyHttpMessages,
   GenericStaticResponse,
+  Subscription,
+  CyHttpMessages,
 } from './external-types'
 
 export type FixtureOpts = {
@@ -26,9 +27,13 @@ export const SERIALIZABLE_RES_PROPS = _.concat(
   SERIALIZABLE_REQ_PROPS,
   'statusCode',
   'statusMessage',
+  'delay',
+  'throttleKbps',
 )
 
-export const DICT_STRING_MATCHER_FIELDS = ['headers', 'query']
+export const PLAIN_FIELDS: (keyof RouteMatcherOptionsGeneric<any>)[] = ['https', 'port', 'middleware']
+
+export const DICT_STRING_MATCHER_FIELDS: (keyof RouteMatcherOptionsGeneric<any>)[] = ['headers', 'query']
 
 export const STRING_MATCHER_FIELDS = ['auth.username', 'auth.password', 'hostname', 'method', 'path', 'pathname', 'url']
 
@@ -45,56 +50,48 @@ export interface AnnotatedStringMatcher {
  */
 export type AnnotatedRouteMatcherOptions = RouteMatcherOptionsGeneric<AnnotatedStringMatcher>
 
-/** Types for messages between driver and server */
-
-export declare namespace NetEventFrames {
-  export interface AddRoute {
-    routeMatcher: AnnotatedRouteMatcherOptions
-    staticResponse?: BackendStaticResponse
-    hasInterceptor: boolean
-    handlerId?: string
-  }
-
-  interface BaseHttp {
+export declare namespace NetEvent {
+  export interface Http {
     requestId: string
-    routeHandlerId: string
   }
 
-  // fired when HTTP proxy receives headers + body of request
-  export interface HttpRequestReceived extends BaseHttp {
-    req: CyHttpMessages.IncomingRequest
-    /**
-     * Is the proxy expecting the driver to send `HttpRequestContinue`?
-     */
-    notificationOnly: boolean
+  export namespace ToDriver {
+    export interface Event<D> extends Http {
+      subscription: Subscription
+      eventId: string
+      data: D
+    }
+
+    export interface Request extends Event<CyHttpMessages.IncomingRequest> {}
+
+    export interface Response extends Event<CyHttpMessages.IncomingResponse> {}
   }
 
-  // fired when driver is done modifying request and wishes to pass control back to the proxy
-  export interface HttpRequestContinue extends BaseHttp {
-    req: CyHttpMessages.IncomingRequest
-    staticResponse?: BackendStaticResponse
-    hasResponseHandler?: boolean
-    tryNextRoute?: boolean
-  }
+  export namespace ToServer {
+    export interface AddRoute {
+      routeMatcher: AnnotatedRouteMatcherOptions
+      staticResponse?: BackendStaticResponse
+      hasInterceptor: boolean
+      routeId: string
+    }
 
-  // fired when a response is received and the driver has a req.reply callback registered
-  export interface HttpResponseReceived extends BaseHttp {
-    res: CyHttpMessages.IncomingResponse
-  }
+    export interface Subscribe {
+      requestId: string
+      subscription: Subscription
+    }
 
-  // fired when driver is done modifying response or driver callback completes,
-  // passes control back to proxy
-  export interface HttpResponseContinue extends BaseHttp {
-    res?: CyHttpMessages.IncomingResponse
-    staticResponse?: BackendStaticResponse
-    // Millisecond timestamp for when the response should continue
-    delay?: number
-    throttleKbps?: number
-    followRedirect?: boolean
-  }
+    export interface EventHandlerResolved {
+      eventId: string
+      changedData: any
+      /**
+       * If `true`, no further handlers for this event will be called.
+       */
+      stopPropagation: boolean
+    }
 
-  // fired when a response has been sent completely by the server to an intercepted request
-  export interface HttpRequestComplete extends BaseHttp {
-    error?: Error & { code?: string }
+    export interface SendStaticResponse {
+      requestId: string
+      staticResponse: BackendStaticResponse
+    }
   }
 }