+++ packages/runner-ct/src/errors/automation-disconnected.jsx
@@ -6,10 +6,11 @@ export default ({ onReload }) => (
       <p>Whoops, the Cypress extension has disconnected.</p>
       <p className='muted'>Cypress cannot run tests without this extension.</p>
       <button onClick={onReload}>
-        <i className='fas fa-sync-alt'></i> Reload the Browser
+        <i className='fas fa-sync-alt'></i>
+        {' Reload the Browser'}
       </button>
       <div className='helper-line'>
-        <a href='https://on.cypress.io/launching-browsers' target='_blank'>
+        <a href='https://on.cypress.io/launching-browsers' target='_blank' rel='noopener noreferrer'>
           <i className='fas fa-question-circle'></i>
           Why am I seeing this message?
         </a>