+++ packages/runner-ct/src/selector-playground/selector-playground.jsx
@@ -36,7 +36,8 @@ class SelectorPlayground extends Component {
       <div className={cs('header-popup selector-playground', `method-${model.method}`, {
         'no-elements': !model.numElements,
         'invalid-selector': !model.isValid,
-      })}>
+      })}
+      >
         <div className='selector'>
           <Tooltip
             title='Click an element to see a suggested selector'
@@ -44,7 +45,8 @@ class SelectorPlayground extends Component {
           >
             <button
               className={`highlight-toggle ${model.isEnabled ? 'active' : ''}`}
-              onClick={this._toggleEnablingSelectorPlayground}>
+              onClick={this._toggleEnablingSelectorPlayground}
+            >
               <span className='fa-stack'>
                 <i className='far fa-square fa-stack-1x'></i>
                 <i className='fas fa-mouse-pointer fa-stack-1x'></i>
@@ -57,7 +59,9 @@ class SelectorPlayground extends Component {
           >
             {this._methodSelector()}
             <span>(</span>
-            <span>{'\''}</span>
+            <span>
+              {'\''}
+            </span>
             <div className='selector-input'>
               <input
                 ref={(node) => this._input = node}
@@ -67,15 +71,17 @@ class SelectorPlayground extends Component {
                 onFocus={this._setHighlight(true)}
               />
             </div>
-            <span>{'\''}</span>
+            <span>
+              {'\''}
+            </span>
             <span>)</span>
-            <input ref='copyText' className='copy-backer' value={selectorText} readOnly />
+            {/* eslint-disable-next-line react/no-string-refs */}
+            <input ref='copyText' className='copy-backer' value={selectorText} readOnly={true} />
             <Tooltip title={model.infoHelp || ''} className='cy-tooltip'>
               <span className='info num-elements'>
                 {model.isValid ?
                   model.numElements :
-                  <i className='fas fa-exclamation-triangle'></i>
-                }
+                  <i className='fas fa-exclamation-triangle' />}
               </span>
             </Tooltip>
           </div>
@@ -83,8 +89,8 @@ class SelectorPlayground extends Component {
             <button
               ref={(node) => this._copyButton = node}
               className='copy-to-clipboard'
-              onClick={this._copyToClipboard}
               disabled={!model.numElements || !model.isValid}
+              onClick={this._copyToClipboard}
               onMouseOut={fixMouseOut(this._resetCopyText, () => this._copyButton)}
             >
               <i className='far fa-copy' />
@@ -94,19 +100,21 @@ class SelectorPlayground extends Component {
             <button
               ref={(node) => this._printButton = node}
               className='print-to-console'
-              onClick={this._printToConsole}
               disabled={!model.numElements || !model.isValid}
+              onClick={this._printToConsole}
               onMouseOut={fixMouseOut(this._resetPrintText, () => this._printButton)}
             >
               <i className='fas fa-terminal' />
             </button>
           </Tooltip>
         </div>
-        <a className='selector-info' href='https://on.cypress.io/selector-playground' target="_blank">
-          <i className='fas fa-question-circle'></i>{' '}
-          Learn more
+        <a className='selector-info' href='https://on.cypress.io/selector-playground' target="_blank" rel="noreferrer">
+          <i className='fas fa-question-circle'></i>
+          {' Learn more'}
         </a>
-        <button className='close' onClick={this._togglePlaygroundOpen}>x</button>
+        <button className='close' onClick={this._togglePlaygroundOpen}>
+x
+        </button>
       </div>
     )
   }
@@ -143,15 +151,16 @@ class SelectorPlayground extends Component {
     return (
       <span className={cs('method', {
         'is-showing': this.showingMethodPicker,
-      })}>
+      })}
+      >
         <button onClick={this._toggleMethodPicker}>
-          <i className='fas fa-caret-down'></i>{' '}
-          cy.{model.method}
+          <i className='fas fa-caret-down'></i>
+          {` cy. ${model.method}`}
         </button>
         <div className='method-picker'>
           {_.map(methods, (method) => (
             <div key={method} onClick={() => this._setMethod(method)}>
-              cy.{method}
+              {`cy.${method}`}
             </div>
           ))}
         </div>
@@ -188,6 +197,7 @@ class SelectorPlayground extends Component {
 
   _copyToClipboard = () => {
     try {
+      // eslint-disable-next-line react/no-string-refs
       this.refs.copyText.select()
       const successful = document.execCommand('copy')