+++ packages/runner/src/iframe/iframes.jsx
@@ -149,15 +149,6 @@ export default class Iframes extends Component {
 
     this.autIframe.showInitialBlankContents()
 
-    // specs with type "component" can only arrive if the server has "componentTesting" experiment on
-    if (this.props.config.spec.specType === 'component') {
-      // In mount mode we need to render something right from spec file
-      // So load application tests to the aut frame
-      $autIframe.prop('src', specSrc)
-
-      return $autIframe
-    }
-
     const $specIframe = $('<iframe />', {
       id: `Your Spec: '${specSrc}'`,
       class: 'spec-iframe',